move create to block_reader.rs
